requirements.txt
crewai>=0.28.0
crewai-tools>=0.4.0
langchain-openai>=0.1.0
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pycaret[full]>=3.2.0
shap>=0.44.0
lime>=0.2.0
python-dotenv>=1.0.0